Minor fixes (#76)
* Add .dockerignore files for client, database, and server directories

* Update action permissions to include moderators for archiving reservations

* Add client role option to user registration form

* Bump version to 0.4.1-alpha in package.json, package-lock.json, Dockerfile, and build.gradle; update Docker images in docker-compose.yml and DrawerMenu component

* Wrap EmployeesStatusWidget in AdminOnly component for access control

* Add immune property to User model and update deletion logic in UserController

* Add password_reset_token and immune fields to master administrator insert statement

+++ server/src/main/java/com/kalyvianakis/estiator/io/controller/UserController.java
@@ -19,6 +19,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.MethodNotAllowedException;
 
 import java.sql.Time;
 import java.time.LocalDate;
@@ -169,6 +170,11 @@ public class UserController {
     public ResponseEntity<?> delete(@PathVariable Long id) throws ResourceNotFoundException, IllegalArgumentException {
         if(userService.notExists(id)) {
             throw new ResourceNotFoundException("User not found for ID: " + id);
+        }
+
+        User user = userService.get(id);
+        if (user.isImmune()) {
+            throw new IllegalArgumentException("User cannot be deleted for ID: " + id);
         }
 
         userService.delete(id);
